backend/requirements.txt
fastapi
uvicorn
motor
pyjwt
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pymongo[srv]